pyproject.toml
[project]
name = "tiebacontentreviewer"
version = "0.1.0"
description = "A content reviewer service for Tiebameow."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "orjson>=3.11.5",
    "protobuf>=6.33.2",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "tiebameow[orm]>=0.4.1,<0.5.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "src.rpc.*"
ignore_errors = true

[tool.pytest.ini_options]
addopts = "--strict-config --strict-markers --cov=src --cov-report=term-missing"
asyncio_mode = "strict"

[tool.coverage.run]
omit = [
    "src/rpc/*",
]
